Add alert

+++ src/rete/editor.js
@@ -188,20 +188,6 @@ const init = async () => {
 }
 
 export const exportEditorData = () => {
-	var FILE_KEY;
-	function addKey() {
-		var list = localStorage.getItem('List');
-		if (list !== null){
-			var array = list.split(',');
-			if (!array.includes(FILE_KEY)){
-				localStorage.setItem('List',list+FILE_KEY+',');
-			}
-		} else {
-			localStorage.setItem('List',FILE_KEY+',');
-		}
-
-
-	}
 
 	function retrieveSave() {
 		let editorData = editor.toJSON();
@@ -229,16 +215,20 @@ export const exportEditorData = () => {
 
 	});
 	// Renvoyer le resultat de la lecture du fichier sous forme txt
-	reader.readAsText(file);
-	element.href = URL.createObjectURL(file);
-	element.download = debug.file.nodes[1].data.titre + "_-_Story_file.json";
-	console.log(debug)
-	FILE_KEY=element.download;
-	localStorage.setItem(FILE_KEY,JSON.stringify(debug));
-	console.log(debug)
-	addKey();
-	document.body.appendChild(element); // Required for this to work in FireFox
-	element.click();
+	if (debug.file.nodes[1].data.titre == ""){
+		Swal.fire({
+			icon: 'error',
+			title: 'Oops...',
+			text: 'Vous avez oublié de donner un nom à votre histoire !',
+		  })
+	} else {
+		reader.readAsText(file);
+		element.href = URL.createObjectURL(file);
+		element.download = debug.file.nodes[1].data.titre + "_-_Story_file.json";
+		document.body.appendChild(element); // Required for this to work in FireFox
+		element.click();
+	}
+
 }
 
 export const loadEditorData = (event) => {
@@ -290,19 +280,59 @@ export const saveEditorData = (event) => {
 	// Définie le contenu qui va être dans le fichier JSON
 	var debug = retrieveSave();
 
-	
-	console.log(debug)
-	FILE_KEY=debug.file.nodes[1].data.titre + "_-_Story_file.json";
-	localStorage.setItem(FILE_KEY,JSON.stringify(debug));
-	console.log(debug)
-	addKey();
-	document.body.appendChild(element); // Required for this to work in FireFox
-	element.click();
+	// demande à l'utilisateur de rentrer un titre si le titre est vide !
+	if (debug.file.nodes[1].data.titre == ""){
+		Swal.fire({
+			icon: 'error',
+			title: 'Oops...',
+			text: 'Vous avez oublié de donner un nom à votre histoire !',
+		  })
+	}
+	// Si l'utilisateur a rentré un titre
+	else {
+		console.log(debug)
+		FILE_KEY=debug.file.nodes[1].data.titre + "_-_Story_file.json";
 
-	Swal.fire({
-		icon: 'success',
-		title: 'Votre fichier a bien été sauvegardé',
-		showConfirmButton: false,
-		timer: 1000
-	  })
+		if (localStorage.getItem(FILE_KEY)!== "" && localStorage.getItem(FILE_KEY)!== null){
+			Swal.fire({
+				title: 'Une histoire portant ce nom existe déjà',
+				text: "Voulez vous quand même sauvegarder votre histoire ?",
+				icon: 'warning',
+				showCancelButton: true,
+				confirmButtonColor: '#3085d6',
+				cancelButtonColor: '#d33',
+				confirmButtonText: 'Sauvegarder',
+				cancelButtonText: 'Annuler'
+			  }).then((result) => {
+				if (result.value) {
+					localStorage.setItem(FILE_KEY,JSON.stringify(debug));
+					console.log(debug)
+					addKey();
+					document.body.appendChild(element); // Required for this to work in FireFox
+					element.click();
+				
+					Swal.fire({
+						icon: 'success',
+						title: 'Votre fichier a bien été sauvegardé',
+						showConfirmButton: false,
+						timer: 1000
+					  })
+				}
+			  })
+		} else {
+			localStorage.setItem(FILE_KEY,JSON.stringify(debug));
+			console.log(debug)
+			addKey();
+			document.body.appendChild(element); // Required for this to work in FireFox
+			element.click();
+		
+			Swal.fire({
+				icon: 'success',
+				title: 'Votre fichier a bien été sauvegardé',
+				showConfirmButton: false,
+				timer: 1000
+			  })
+		}
+		}
+
 }
